add lua_const definition with dep on lua_bind_hash crate
